Move entities to Entities sub-folder, hide bin and obj folders in vscode
