+++ Averager.cs
@@ -1,11 +1,9 @@
 using System;
-using System.Collections.Generic;
 using System.Threading.Tasks;
-using Azure;
 using Azure.Data.Tables;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Extensions.Logging;
-using Microsoft.WindowsAzure.Storage.Table;
+using SteveSyrell.PurpleAirUploadApi.Entities;
 
 namespace SteveSyrell.PurpleAirUploadApi
 {
