+++ RealTime.cs
@@ -2,14 +2,13 @@ using System;
 using System.IO;
 using System.Text.Json;
 using System.Threading.Tasks;
-
-using Azure;
 using Azure.Data.Tables;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.Http;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
+using SteveSyrell.PurpleAirUploadApi.Entities;
 
 namespace SteveSyrell.PurpleAirUploadApi
 {
